Add full windows support
This very large patch is the culminating point of a couple of years
getting libfyaml to build and work fully in windows.

There is full support for building on windows using MSVC/clang or GCC.
Native windows builds should work out of the box, using VSCode or
anything else really.

Additionally you can use msvc-wine to install the MSVC redistributables
on linux and compile either using msvc using wine or clang-cl or
even a full clang cross compilation triplet.

Compile using cl using wine
$ cmake -DMSVC_WINE_DIR=/path/to/msvc -DMSVC_WINE_COMPILER=cl ..

Compile using clang-cl without wine
$ cmake -DMSVC_WINE_DIR=/path/to/msvc -DMSVC_WINE_COMPILER=clang-cl ..

Compile using clang cross compile without wine
$ cmake -DMSVC_WINE_DIR=/path/to/msvc -DMSVC_WINE_COMPILER=clang ..

Although this is the first windows supporting release it should
be stable and performant.

Many thanks to all libfyaml contributors that submitted patches.
It would not be possible without them.

+++ src/lib/fy-token.c
@@ -1386,7 +1386,7 @@ fy_token_set_comment(struct fy_token *fyt, enum fy_comment_placement which,
 		if (fy_token_is_lb(fyt, c)) {
 			c = '\n';
 		} else {
-			assert(w <= sizeof(utf8buf));
+			assert((size_t)w <= sizeof(utf8buf));
 			fy_utf8_put_unchecked(utf8buf, c);
 			fwrite(utf8buf, 1, w, fp);
 		}
